+++ server.js
@@ -14,4 +14,4 @@ app.get('*', (req, res) => {
 const PORT = process.env.PORT || 3001;
 app.listen(PORT, () => {
   console.log(`http://localhost:${PORT}`);
-});
+});
