--- src/index.css
@@ -1,12 +1,0 @@
-*{
-  margin: 0;
-  padding: 0;
-}
-
-.MainPage-Contents {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-  }
