+++ src/style/Mainpage.css
@@ -1,0 +1,12 @@
+*{
+    margin: 0;
+    padding: 0;
+  }
+  
+  .MainPage-Contents {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      height: 100vh;
+    }
